db_test_app/requirements.txt
asgiref==3.3.4
Django==3.2
psycopg2-binary==2.8.5
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
